+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-## 1.9.1
+## Next
+### Features
+* new api for formatters Array with 2 elements regex/string and replacement string
+### Bugs
 * fix cursor in IE and iOS/Safari reported by @RinaVladimyrovna [#350](https://github.com/jcubic/jquery.terminal/issues/350)
 * don't apply formatters in echo commands for completion (found by applying completion to emoji demo)
 * fix substring and html entity (entering < & > in command line was showing entity not character)